clipos-data/clipos-hardware: Add i8042 module to kvm_ovmf64 profile
This module is needed to detect AT keyboard and PS/2 mouse before atkbd
and psmouse can take over, but neither of both these resulting modules
depend on it. Therefore we need to add it to the list of explicitly
loaded modules, otherwise module loading can be disabled before the
keyboard has been detected (via PNP) and thus before i8042 was
automatically loaded (resulting in frozen keyboard which in
turn can be quite annoying).

Change-Id: I0385b88dcc61d93abebcdb90f174f887fe0932b6