clipos-meta/clipos-core: Removing unneeded runtime dependency packages
GNU Awk and sed are not requisites in the CLIP OS core for the SDK
configure step (Awk and sed from the SDK are used instead)

binutils is utterly unwanted in the CLIP OS core as a
runtime-dependency. It might be one day only in instrumented build
mode if we would like to do compilations on instrumented CLIP OS
instances (but it is not the case yet).

Libraries are not part of the @system set, let package dependency tree
unwinding pull them if they appear to be required for runtime.

Change-Id: Ice4370770c6bf43d216aceaf2af81bc94c5ac9b6
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>