Instrumentation revamp
Switch from levels to features for build instrumentation. Profile
composition is created at SDK startup based on features selected in
intrumentation.toml and given to the SDk by cosmk.

In order not to collide with upstream Gentoo USE flags, we make use of
the USE_EXPAND feature of Portage to create some kind of USE flag
namespaces for USE flags related to the CLIP OS project only.

Change-Id: I84d00b8af96ca6ddfc012d981970915ba6f26f25
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>