clipos-product-info: Rework to drop products dependency
We do not want any dependency from the CLIP OS (or derivative) recipe
and product definitions into the Portage tree.

The "clipos-product-info" eclass was a way to introduce such a
dependency. Therefore we deliberately hard code product values in this
eclass.

Derivatives will be able to override those values by creating a similar
eclass in their overlay.

Change-Id: I8aae9966badaa3c42fabfeb2f78cd9b3f639f01a
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>