profiles: Introduce CLIPOS_USE USE_EXPAND
"USE flag namespace" which will hold USE flags that are of interest but
are not managed by the Gentoo upstream project. Some of the USE flags
fitting in this namespace may be suggested to the Gentoo project for
upstream integration. The rationale behind the use of a dedicated
"USE_EXPAND" with a "clipos" prefix is to prevent any collision with
potential upcoming USE flags similarly named (and which may have
different meaning than ours).

Change-Id: If1e648a344db0a46feaa9ece87ef98570336b796
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>