profiles: Keep caps & acl useflags enabled for coreutils
ACL and capabilities are enabled in the kernel and supported by the
filesystems. Removing ACL support in the system tools will only hide
them, which is error prone. Thus support is disabled by default but
enabled for coreutils.

Keep ACL and capability support disabled for efiboot as we do not use
them in the initramfs which is also ephemeral.

Change-Id: I652ad539ddd9690463ee23a94c77ce2c53a83dbc
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>